claude_tests/test_ecommercev1/requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
lifetimes>=0.11.3
tabulate>=0.9.0
